MacOS: get refresh rate

+++ src/Platform/IPlatform.hpp
@@ -8,7 +8,7 @@ struct IPlatform
 	virtual ~IPlatform() = default;
 	virtual void setIcon(const sf::WindowHandle& inHandle) = 0;
 	virtual void toggleFullscreen(const sf::WindowHandle& inHandle, const sf::Uint32 inStyle, const bool inWindowed, const sf::Vector2u& inResolution) = 0;
-	virtual int getRefreshRate(const sf::WindowHandle& inHandle) = 0;
+	virtual float getRefreshRate(const sf::WindowHandle& inHandle) = 0;
 	virtual float getScreenScalingFactor(const sf::WindowHandle& inHandle) = 0;
 };
 }
