+++ src/Platform/Unix/LinuxPlatform.hpp
@@ -12,7 +12,7 @@ struct LinuxPlatform : IPlatform
 	void setIcon(const sf::WindowHandle& inHandle) final;
 	void toggleFullscreen(const sf::WindowHandle& inHandle, const sf::Uint32 inStyle, const bool inWindowed, const sf::Vector2u& inResolution) final;
 	float getScreenScalingFactor(const sf::WindowHandle& inHandle) final;
-	int getRefreshRate(const sf::WindowHandle& inHandle) final;
+	float getRefreshRate(const sf::WindowHandle& inHandle) final;
 };
 }
